+++ src/spec/EngineSpec.ts
@@ -57,4 +57,49 @@ describe('The engine', () => {
       expect(fired).toBe(true);
    });
 
+   it('should emit a predraw event', () => {
+      var fired = false;
+      engine.on('predraw', () => fired = true);
+
+      loop.advance(100);
+
+      expect(fired).toBe(true);
+   });
+
+   it ('should emit a postdraw event', () => {
+      var fired = false;
+      engine.on('postdraw', () => fired = true);
+
+      loop.advance(100);
+
+      expect(fired).toBe(true);
+   });
+   
+   it('should tell engine is running', () => {
+      var status = engine.isPaused();
+      expect(status).toBe(false);
+   });
+   
+   it('should tell engine is paused', () => {
+      engine.stop();
+      var status = engine.isPaused();
+      expect(status).toBe(true);
+   });
+   
+   it('should again tell engine is running', () => {
+      engine.start();
+      var status = engine.isPaused();
+      expect(status).toBe(false);
+   });
+
+   it('should return screen dimensions', () => {
+      engine.start();
+      var left = engine.screenToWorldCoordinates(ex.Vector.Zero).x;
+      var top = engine.screenToWorldCoordinates(ex.Vector.Zero).y;
+      var right = left + engine.getDrawWidth();
+      var bottom = top + engine.getDrawHeight();
+      var localBoundingBox = new ex.BoundingBox(left, top, right, bottom);
+      expect(engine.getWorldBounds()).toEqual(localBoundingBox);
+   });
+   
 });